Compute the index of Divergent Evolutionary Trajectories.

The index of Divergent Evolutionary Trajectories is a measure derived from Shannon's entropy to determine, for any driver event X, how heterogeneous are the trajectories that lead to X.

This is simply based on counting the number of edges Y -> X, for every Y in the data, and using the distribution of the observed frequencies to determine an entropy-derived measure (analog to species heterogeneity). This means that for values larger than 0, the model observers heterogeneous trajectories in the data.

To compute the DET the cohort must have fits available.

DET_index(x, drivers = x$variantIDs.driver, min.occurrences = 0)

Arguments

x
A REVOLVER cohort object with fits available.

drivers
The list of drivers to compute the DET for.

min.occurrences
The minimum number of occurrences for a trajectory to be considered, zero by default.

Value

The DET index for the input cohort.

See also

Examples

# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the DET with all cohort
DET_index(TRACERx_NEJM_2017_REVOLVER)

#> # A tibble: 79 x 4
#>    driver diversity     N DET_index
#>    <chr>      <dbl> <int>     <dbl>
#>  1 APC            0     1         0
#>  2 BRAF           0     1         0
#>  3 CDKN2A         0     1         0
#>  4 CHEK2          0     1         0
#>  5 CIC            0     1         0
#>  6 CMTR2          0     1         0
#>  7 COL2A1         0     1         0
#>  8 CREBBP         0     1         0
#>  9 CUX1           0     1         0
#> 10 DICER1         0     1         0
#> # … with 69 more rows

# Look specifically for TP53 - the DET suggests
# heterogeneous trajectories.
DET_index(TRACERx_NEJM_2017_REVOLVER, drivers = 'TP53')

#> # A tibble: 1 x 4
#>   driver diversity     N DET_index
#>   <chr>      <dbl> <int>     <dbl>
#> 1 TP53        1.63     8     0.786

# Look specifically for drivers in at least 5 patients
DET_index(TRACERx_NEJM_2017_REVOLVER, min.occurrences = 5)

#> # A tibble: 17 x 4
#>    driver diversity     N DET_index
#>    <chr>      <dbl> <int>     <dbl>
#>  1 CDKN2A     0         1     0    
#>  2 COL2A1     0         1     0    
#>  3 EGFR       0         1     0    
#>  4 FAT1       0         1     0    
#>  5 FGFR1      0         1     0    
#>  6 KEAP1      0         1     0    
#>  7 MYC        0         1     0    
#>  8 NFE2L2     0         1     0    
#>  9 NOTCH1     0         1     0    
#> 10 PIK3CA     0         1     0    
#> 11 PTEN       0         1     0    
#> 12 RASA1      0         1     0    
#> 13 SOX2       0         1     0    
#> 14 STK11      0         1     0    
#> 15 TERT       0         1     0    
#> 16 KRAS       0.598     2     0.863
#> 17 TP53       1.42      5     0.880