Extract all data for a patient.

From a cohort object, this function extracts the tibble of all data available for a patient (either CCF or binary).

Data(x, p)

Arguments

x
A REVOLVER cohort.

p
The id of a patient in the cohort.

Value

Drivers data for a custom patient.

See also

Examples

# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

Data(TRACERx_NEJM_2017_REVOLVER, 'CRUK0002')

#> # A tibble: 7 x 12
#>   id    Misc  patientID variantID is.driver is.clonal cluster cluster_size CCF  
#>   <chr> <chr> <chr>     <chr>     <lgl>     <lgl>     <chr>          <int> <chr>
#> 1 __mu… CRUK… CRUK0002  RB1       TRUE      FALSE     1                  3 R1:0…
#> 2 __mu… CRUK… CRUK0002  IKZF1     TRUE      FALSE     1                  3 R1:0…
#> 3 __mu… Anno… CRUK0002  KRAS      TRUE      FALSE     1                  3 R1:0…
#> 4 __mu… CRUK… CRUK0002  MET       TRUE      TRUE      2                  2 R1:0…
#> 5 __mu… Anno… CRUK0002  TERT      TRUE      TRUE      2                  2 R1:0…
#> 6 __mu… CRUK… CRUK0002  NF1       TRUE      FALSE     5                  1 R1:0…
#> 7 __mu… CRUK… CRUK0002  EP300     TRUE      FALSE     6                  1 R1:0…
#> # … with 3 more variables: R1 <dbl>, R2 <dbl>, R3 <dbl>

Data(TRACERx_NEJM_2017_REVOLVER, 'CRUK0008')

#> # A tibble: 6 x 11
#>   id    Misc  patientID variantID is.driver is.clonal cluster cluster_size CCF  
#>   <chr> <chr> <chr>     <chr>     <lgl>     <lgl>     <chr>          <int> <chr>
#> 1 __mu… CRUK… CRUK0008  KEAP1     TRUE      TRUE      2                  5 R1:0…
#> 2 __mu… CRUK… CRUK0008  STK11     TRUE      TRUE      2                  5 R1:0…
#> 3 __mu… CRUK… CRUK0008  PRDM1     TRUE      TRUE      2                  5 R1:0…
#> 4 __mu… CRUK… CRUK0008  U2AF1     TRUE      TRUE      2                  5 R1:0…
#> 5 __mu… Anno… CRUK0008  MYC       TRUE      TRUE      2                  5 R1:0…
#> 6 __mu… CRUK… CRUK0008  ARID2     TRUE      FALSE     3                  1 R1:0…
#> # … with 2 more variables: R1 <dbl>, R2 <dbl>