Palette of n distinct colours.

ColorRamp palette extrapolating all qualitative (qual) palettes from the RColorBrewer package. This function can have the number of required colours as input parameter; if this does not happen a function is returned.

distinct_palette_many(n = NULL)

Arguments

n
The number of required colours. By default this is NULL.

Value

Either a function that applied to a number returns the required colours, or the actual sampling function.

See also

Examples

distinct_palette_many(2)

#> [1] "#7FC97F" "#FFED6F"