Compute/ add CCF-based phylogenies to a REVOLVER cohort.

Compute/ add CCF-based phylogenies to a REVOLVER cohort.

revolver_compute_phylogenies(x, patient, precomputed.trees = NULL,
  precomputed.scores = NULL, options = list(sspace.cutoff = 10000,
  n.sampling = 5000, overwrite = FALSE, store.max = 100),
  verbose = FALSE)

Arguments

x
An object of class "rev_cohort"

patient
A patient in the cohort, for which phylogenies are created

precomputed.trees
If a list a of precomputed trees is available, this list should contain their adjacency matrix. No computation will be carried out in this case.

precomputed.scores
If a list a of precomputed trees is available, this list should contain their scores. No computation will be carried out in this case.

options
If one wants to generate new phylogenies for this patient, there named parameters should be passed through this list. sspace.cutoff = 10000 is the state-space cutoff to generate trees in an exhaustive fashion, or to Montecarlo sample them. If Montecarlo is chosen sampling tells how many trees are sampled. overwrite = FALSE sets if the patient's current phylogenies should be overwritten, in case they are already available. store.max = 100 tells how many trees should be stored, if there are more than store.max available (ranked).

verbose
output type.

Value

a modififed object of class "rev_cohort" with available phylogeneis for patient.

Details

Create or compute phylogenies for REVOLVER fit from CCF data. This method should be used also if one has already pre-computed trees to use for fitting.

Examples

# NOT RUN {
 TODO
# }