REVOLVER constructor for a tree

This constructor creates an object of type "rev_phylo", which represents a model for a patient. This is the same type of object regarldess data are CCF or binary values.

revolver_phylogeny(x, patient, M, score,
  annotation = paste0("A tree for patient ", patient))

Arguments

x
A REVOLVER cohort.

patient
The id of this patient in the cohort.

M
The tree adjacency matrix.

score
A score for this model, that we seek to maximize.

annotation
A string to annotate this tree.

Value

An object of class "rev_phylo" that represents this tree.

Note

This function is meant to be called after a REVOLVER cohort object has been created, via the constructure revolver_cohort.

Examples

# We re-build the trees for one patient
data(TRACERx_cohort)

# Because the function will return a new cohort, we can just
M = ...

#> Error in eval(expr, envir, enclos): '...' used in an incorrect context

# Create a new tree, the score is made up
new_tree = revolver_phylogeny(TRACERx_cohort, patient = "CRUK0002", M, score = 17)

# S3 functions for this object
print(new_tree)

#> function(x,
#>                               patient,
#>                               M,
#>                               score,
#>                               annotation = paste0("A tree for patient ", patient))
#> {
#>   # This function will create this output object
#>   obj <-
#>     structure(
#>       list(
#>         adj_mat = NA,
#>         # Adjacency matrix for the tree
#>         tb_adj_mat = NA,
#>         # Tidygraph object for this tree
#>         score = NA,
#>         # Tree score
#>         patient = NA,
#>         # Patient ID
#>         samples = NA,
#>         # Samples name
#>         drivers = NA,
#>         # Driver mapping to clusters
#>         CCF = NA,
#>         # CCF (aggregated per cluster)
#>         annotation = NA,
#>         # Some custom annotation
#>         transfer = NA,
#>         # Information transfer from this tree
#>         tree_type = NA       # Mutation tree or phylogeny
#>       ),
#>       class = "rev_phylo",
#>       call = match.call()
#>     )
#>   
#>   # =-=-=-=-==-=-=-=-=-=-=-=-=-=-=-=-
#>   # The information that we need for each tree are
#>   # CCF clusters, data and driver events (mapped to clusters)
#>   # =-=-=-=-==-=-=-=-=-=-=-=-=-=-=-=-
#>   obj$CCF = CCF_clusters(x, patient)
#>   obj$samples = Samples(x, patient)
#>   obj$drivers = Drivers(x, patient) %>%
#>     select(variantID, cluster)
#>   
#>   # =-=-=-=-==-=-=-=-=-=-=-=-=-=-=-=-
#>   # We begin to create a representation of the tree
#>   # =-=-=-=-==-=-=-=-=-=-=-=-=-=-=-=-
#>   
#>   # Input can should be an adjacency matrix (which works also for monoclonal tumours)
#>   if (class(M) != 'matrix')
#>     stop("Input `M` should be an adjacency matrix, aborting.")
#>   
#>   adj_mat =  M
#>   df_mat = MatrixToDataFrame(M)
#>   
#>   # We chack for that to be a tree - empty is OK in this casem if monoclonal
#>   if (!is_tree(adj_mat, allow.empty = nrow(obj$CCF) == 1))
#>     stop("The input adjacency matrix is not a valid tree, aborting.")
#>   
#>   # Add GL node, beware of the special case of empty adj_mat (might happen for a monoclonal tumour)
#>   M_root = ifelse(sum(adj_mat) == 0, rownames(adj_mat), root(adj_mat))
#>   df_mat = rbind(df_mat,
#>                  data.frame(
#>                    from = 'GL',
#>                    to = M_root,
#>                    stringsAsFactors = FALSE
#>                  ))
#>   
#>   # Update the adj matrix with GL, which can now go into obj
#>   obj$adj_mat = DataFrameToMatrix(df_mat)
#>   
#>   # =-=-=-=-==-=-=-=-=-=-=-=-=-=-=-=-
#>   # Create a tidygraph object for this tree
#>   # =-=-=-=-==-=-=-=-=-=-=-=-=-=-=-=-
#>   tb_adj_mat = as_tbl_graph(df_mat) %>%
#>     activate(nodes) %>%
#>     rename(cluster = name)
#>   
#>   # We can add information specific for this tree to the tidygraph
#>   tb_adj_mat = tb_adj_mat %>%
#>     left_join(obj$CCF, by = 'cluster')
#>   
#>   # Sample attachment for input data
#>   attachment = obj$CCF %>%
#>     select(cluster, obj$samples) %>%
#>     reshape2::melt(id = 'cluster') %>%
#>     mutate(value = ifelse(value > 0, 1, 0)) %>%
#>     group_by(variable) %>%
#>     filter(sum(value) == 1) %>%
#>     ungroup() %>%
#>     filter(value == 1) %>%
#>     select(-value) %>%
#>     rename(attachment = variable) %>%
#>     mutate(attachment = paste(attachment)) %>%
#>     group_by(cluster) %>%
#>     summarise(attachment = paste(attachment, collapse = ', '))
#>   
#>   tb_adj_mat = tb_adj_mat %>%
#>     left_join(attachment, by = 'cluster')
#>   
#>   # Drivers per node
#>   tb_adj_mat = tb_adj_mat %>%
#>     left_join(obj$drivers %>%
#>                 group_by(cluster) %>%
#>                 summarise(driver = paste(variantID, collapse = ', ')),
#>               by = 'cluster')
#>   
#>   # Store it in obj
#>   obj$tb_adj_mat = tb_adj_mat
#>   
#>   # Compute the information transfer
#>   obj$transfer = information_transfer(obj)
#>   
#>   # Extras
#>   obj$score = score
#>   obj$patient = patient
#>   
#>   obj$annotation = annotation
#>   
#>   obj$tree_type = ifelse(
#>     all(obj$CCF %>% select(obj$samples) %in% c(0, 1)),
#>     "Mutation tree (binary data)",
#>     "Phylogenetic tree (CCF)"
#>   )
#>   
#>   return(obj)
#> }
#> <environment: 0x7fd992272d48>

plot(new_tree)

#> Error in x$CCF: $ operator is invalid for atomic vectors