Remove a driver event from a cohort.

Remove a driver event from a cohort.

revolver_removeDriver(x, patient, variantID, misc, cluster)

Arguments

x
An object of class "rev_cohort"

patient
Patient ID.

variantID
ID of the event to mark as FALSE.

misc
misc of the event to mark as FALSE

cluster
cluster of the event to mark as FALSE

Value

a modified cohort object of class "rev_cohort"

Details

Basic editing function. Each event is identied through its ID, cluster assignment and its misc flag. With this function, you can remove it. The event is not physically removed from the dataset, but instead its flag is.Driver is set to FALSE. If there are phylogenies inside, they are updated as well. If you have fit the models, however, you should re-run the fit after this modification because this modification does not propagate.

Examples

data(CRC.cohort)

#> Warning: data set ‘CRC.cohort’ not found

head(CRC.cohort$dataset) # Info to use to remove

#> Error in head(CRC.cohort$dataset): object 'CRC.cohort' not found

new.data = revolver_removeDriver(CRC.cohort, "adenoma_1", "APC", "NOTHING", "1")

#>  [ REVOLVER remove driver across all samples ] 
#> 	 -Patient adenoma_1 
#> 	 -Alteration ID APC 
#> 	 -Misc NOTHING 
#> 	 -Group/ cluster 1 
#> 

#> Error in x$dataset[x$dataset$is.driver & x$dataset$patientID == patient &     x$dataset$Misc == misc & x$dataset$cluster == cluster, "is.driver"] = FALSE: object 'CRC.cohort' not found

head(new.data$dataset)

#> Error in head(new.data$dataset): object 'new.data' not found