Summary statistics for the cohort's patients.

Returns the number of samples per patient, the number of drivers, the number of clonal and subclonal mutations etc. The function can be run on a subset of patients.

Stats(x, patients = x$patients)

Arguments

x
A REVOLVER cohort.

patients
The patients for which the summaries are computed.

Value

A tibble with summary stastics.

See also

Examples

# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the stats for all patients
Stats(TRACERx_NEJM_2017_REVOLVER)

#> # A tibble: 99 x 7
#>    patientID numBiopsies numMutations numDriverMutati… numClonesWithDr…
#>    <chr>           <int>        <int>            <int>            <int>
#>  1 CRUK0001            3            7                7                4
#>  2 CRUK0002            3            7                7                4
#>  3 CRUK0003            5            4                4                2
#>  4 CRUK0004            4            4                4                2
#>  5 CRUK0005            4            6                6                2
#>  6 CRUK0006            2            6                6                3
#>  7 CRUK0007            2            3                3                1
#>  8 CRUK0008            2            6                6                2
#>  9 CRUK0009            4            7                7                2
#> 10 CRUK0010            2            3                3                1
#> # … with 89 more rows, and 2 more variables: numTruncalMutations <int>,
#> #   numSubclonalMutations <int>

# And subset the patients
Stats(TRACERx_NEJM_2017_REVOLVER, patients = c('CRUK0001', 'CRUK0002'))

#> # A tibble: 2 x 7
#>   patientID numBiopsies numMutations numDriverMutati… numClonesWithDr…
#>   <chr>           <int>        <int>            <int>            <int>
#> 1 CRUK0001            3            7                7                4
#> 2 CRUK0002            3            7                7                4
#> # … with 2 more variables: numTruncalMutations <int>,
#> #   numSubclonalMutations <int>