Summary statistics for the cohort's trees.

Returns the statistics about the trees that are available in the cohort. The function can be run on a subset of patients.

Stats_trees(x, patients = x$patients)

Arguments

x

A REVOLVER cohort.

patients

The patients for which the summaries are computed.

Value

A tibble with the driver stastics.

See also

Examples

# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

# Get the stats for all patients
Stats_trees(TRACERx_NEJM_2017_REVOLVER)
#> # A tibble: 99 × 6
#>    patientID hasTrees numTrees maxScore minScore combInfTransf
#>    <chr>     <lgl>       <int>    <dbl>    <dbl>         <int>
#>  1 CRUK0001  TRUE            3    0.111   0.111              3
#>  2 CRUK0002  TRUE            2    0.75    0.0833             2
#>  3 CRUK0003  TRUE            1    1       1                  1
#>  4 CRUK0004  TRUE            1    1       1                  1
#>  5 CRUK0005  TRUE            1    1       1                  1
#>  6 CRUK0006  TRUE            2    0.667   0.167              2
#>  7 CRUK0007  TRUE            1    1       1                  1
#>  8 CRUK0008  TRUE            1    1       1                  1
#>  9 CRUK0009  TRUE            1    1       1                  1
#> 10 CRUK0010  TRUE            1    1       1                  1
#> # … with 89 more rows

# And subset the patients
Stats_trees(TRACERx_NEJM_2017_REVOLVER, patients = c('CRUK0001', 'CRUK0002'))
#> # A tibble: 2 × 6
#>   patientID hasTrees numTrees maxScore minScore combInfTransf
#>   <chr>     <lgl>       <int>    <dbl>    <dbl>         <int>
#> 1 CRUK0001  TRUE            3    0.111   0.111              3
#> 2 CRUK0002  TRUE            2    0.75    0.0833             2