Compute binary mutation trees for a patient (phylogenies).

Create mutation tree phylogenies from the package btree, fitting the input data given to REVOLVER. A set of patient ids can be given as input, by default all are used. A parmeter controls if you want to overwrite the trees of the patient, where they already computed. Please refer to the parameters of the btree constructor from package btree for the input parameters that you can use to tune the construction.

compute_mutation_trees(
  x,
  patients = Stats_cohort(x) %>% pull(patientID),
  overwrite = FALSE,
  ...
)

Arguments

x
A REVOLVER cohort with now available phylogeneis for the required patients.

patients
A set of patient ids in the cohort, for which the phylogenies are created.

Note

This function is not yet implemented

See also

Examples

print('TODO')

#> [1] "TODO"