Return summary features for the cohort.

Computes a set of summary features for the cohort, in the form of a matrix. The fits must be available inside the cohort to retrieve some of these features, which are:

the matrix of driver events with mean CCF or binary value;

the matrix of the drivers occurrence acrross all the cohort;

the matrix of the clonal drivers occurrence acrross all the cohort;

the matrix of the subclonal drivers occurrence acrross all the cohort;

the matrix of the occurrence of all evolutionary trajectories across patients

The function returns a named list, so that names can be used to access the matrices.

get_features(x, patients = x$patients)

Arguments

x
A REVOLVER cohort.

patients
A vector of patient ids for which the features are extracted.

Value

A list of matrices.

See also

Examples

# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

features = get_features(TRACERx_NEJM_2017_REVOLVER)

print(names(features))

#> [1] "Matrix_mean_CCF"          "Matrix_drivers"          
#> [3] "Matrix_clonal_drivers"    "Matrix_subclonal_drivers"
#> [5] "Matrix_trajectories"     

print(features)

#> $Matrix_mean_CCF
#> # A tibble: 99 x 80
#>    patientID   APC ARHGAP35 ARID1B ARID2 ASXL1   ATM  BAP1  BRAF  CBLB CCND1
#>    <chr>     <dbl>    <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 CRUK0001      0    0.38       0  0        0     0     0 0         0     0
#>  2 CRUK0002      0    0          0  0        0     0     0 0         0     0
#>  3 CRUK0003      0    0          0  0        0     0     0 0         0     0
#>  4 CRUK0004      0    0          0  0        0     0     0 0         0     0
#>  5 CRUK0005      0    0          0  0        0     0     0 0.99      0     0
#>  6 CRUK0006      0    0          0  0        0     0     0 0         0     0
#>  7 CRUK0007      0    0          0  0        0     0     0 0         0     0
#>  8 CRUK0008      0    0          0  0.43     0     0     0 0         0     0
#>  9 CRUK0009      0    0.988      0  0        0     0     0 0.988     0     0
#> 10 CRUK0010      0    0          0  0        0     0     0 0         0     0
#> # … with 89 more rows, and 69 more variables: CDKN2A <dbl>, CHEK2 <dbl>,
#> #   CIC <dbl>, CMTR2 <dbl>, COL2A1 <dbl>, COL5A2 <dbl>, CREBBP <dbl>,
#> #   CTNNB1 <dbl>, CUX1 <dbl>, CYLD <dbl>, DICER1 <dbl>, DNM2 <dbl>, EGFR <dbl>,
#> #   EP300 <dbl>, FANCC <dbl>, FANCM <dbl>, FAS <dbl>, FAT1 <dbl>, FBXW7 <dbl>,
#> #   FGFR1 <dbl>, FLT4 <dbl>, GATA3 <dbl>, IKZF1 <dbl>, KEAP1 <dbl>,
#> #   KMT2C <dbl>, KMT2D <dbl>, KRAS <dbl>, LATS1 <dbl>, MAP3K1 <dbl>, MET <dbl>,
#> #   MGA <dbl>, MLH1 <dbl>, MYC <dbl>, NCOA6 <dbl>, NCOR1 <dbl>, NF1 <dbl>,
#> #   NFE2L2 <dbl>, NOTCH1 <dbl>, NOTCH2 <dbl>, NRAS <dbl>, PASK <dbl>,
#> #   PDGFRA <dbl>, PHOX2B <dbl>, PIK3CA <dbl>, PLXNB2 <dbl>, POLE <dbl>,
#> #   PRDM1 <dbl>, PRF1 <dbl>, PTEN <dbl>, PTPRC <dbl>, RAD21 <dbl>, RASA1 <dbl>,
#> #   RB1 <dbl>, RNF43 <dbl>, SERPINB13 <dbl>, SETD2 <dbl>, SGK223 <dbl>,
#> #   SMAD4 <dbl>, SMARCA4 <dbl>, SOX2 <dbl>, SPEN <dbl>, STK11 <dbl>,
#> #   TERT <dbl>, TP53 <dbl>, TSC2 <dbl>, U2AF1 <dbl>, UBR5 <dbl>, WRN <dbl>,
#> #   WT1 <dbl>
#> 
#> $Matrix_drivers
#> # A tibble: 99 x 80
#>    patientID   APC ARHGAP35 ARID1B ARID2 ASXL1   ATM  BAP1  BRAF  CBLB CCND1
#>    <chr>     <dbl>    <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 CRUK0001      0        1      0     0     0     0     0     0     0     0
#>  2 CRUK0002      0        0      0     0     0     0     0     0     0     0
#>  3 CRUK0003      0        0      0     0     0     0     0     0     0     0
#>  4 CRUK0004      0        0      0     0     0     0     0     0     0     0
#>  5 CRUK0005      0        0      0     0     0     0     0     1     0     0
#>  6 CRUK0006      0        0      0     0     0     0     0     0     0     0
#>  7 CRUK0007      0        0      0     0     0     0     0     0     0     0
#>  8 CRUK0008      0        0      0     1     0     0     0     0     0     0
#>  9 CRUK0009      0        1      0     0     0     0     0     1     0     0
#> 10 CRUK0010      0        0      0     0     0     0     0     0     0     0
#> # … with 89 more rows, and 69 more variables: CDKN2A <dbl>, CHEK2 <dbl>,
#> #   CIC <dbl>, CMTR2 <dbl>, COL2A1 <dbl>, COL5A2 <dbl>, CREBBP <dbl>,
#> #   CTNNB1 <dbl>, CUX1 <dbl>, CYLD <dbl>, DICER1 <dbl>, DNM2 <dbl>, EGFR <dbl>,
#> #   EP300 <dbl>, FANCC <dbl>, FANCM <dbl>, FAS <dbl>, FAT1 <dbl>, FBXW7 <dbl>,
#> #   FGFR1 <dbl>, FLT4 <dbl>, GATA3 <dbl>, IKZF1 <dbl>, KEAP1 <dbl>,
#> #   KMT2C <dbl>, KMT2D <dbl>, KRAS <dbl>, LATS1 <dbl>, MAP3K1 <dbl>, MET <dbl>,
#> #   MGA <dbl>, MLH1 <dbl>, MYC <dbl>, NCOA6 <dbl>, NCOR1 <dbl>, NF1 <dbl>,
#> #   NFE2L2 <dbl>, NOTCH1 <dbl>, NOTCH2 <dbl>, NRAS <dbl>, PASK <dbl>,
#> #   PDGFRA <dbl>, PHOX2B <dbl>, PIK3CA <dbl>, PLXNB2 <dbl>, POLE <dbl>,
#> #   PRDM1 <dbl>, PRF1 <dbl>, PTEN <dbl>, PTPRC <dbl>, RAD21 <dbl>, RASA1 <dbl>,
#> #   RB1 <dbl>, RNF43 <dbl>, SERPINB13 <dbl>, SETD2 <dbl>, SGK223 <dbl>,
#> #   SMAD4 <dbl>, SMARCA4 <dbl>, SOX2 <dbl>, SPEN <dbl>, STK11 <dbl>,
#> #   TERT <dbl>, TP53 <dbl>, TSC2 <dbl>, U2AF1 <dbl>, UBR5 <dbl>, WRN <dbl>,
#> #   WT1 <dbl>
#> 
#> $Matrix_clonal_drivers
#> # A tibble: 99 x 80
#>    patientID   APC ARHGAP35 ARID1B ARID2 ASXL1   ATM  BAP1  BRAF  CBLB CCND1
#>    <chr>     <dbl>    <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1 CRUK0001      0        0      0     0     0     0     0     0     0     0
#>  2 CRUK0002      0        0      0     0     0     0     0     0     0     0
#>  3 CRUK0003      0        0      0     0     0     0     0     0     0     0
#>  4 CRUK0004      0        0      0     0     0     0     0     0     0     0
#>  5 CRUK0005      0        0      0     0     0     0     0     1     0     0
#>  6 CRUK0006      0        0      0     0     0     0     0     0     0     0
#>  7 CRUK0007      0        0      0     0     0     0     0     0     0     0
#>  8 CRUK0008      0        0      0     0     0     0     0     0     0     0
#>  9 CRUK0009      0        1      0     0     0     0     0     1     0     0
#> 10 CRUK0010      0        0      0     0     0     0     0     0     0     0
#> # … with 89 more rows, and 69 more variables: CDKN2A <dbl>, CHEK2 <dbl>,
#> #   CIC <dbl>, CMTR2 <dbl>, COL2A1 <dbl>, COL5A2 <dbl>, CREBBP <dbl>,
#> #   CTNNB1 <dbl>, CUX1 <dbl>, CYLD <dbl>, DICER1 <dbl>, DNM2 <dbl>, EGFR <dbl>,
#> #   EP300 <dbl>, FANCC <dbl>, FANCM <dbl>, FAS <dbl>, FAT1 <dbl>, FBXW7 <dbl>,
#> #   FGFR1 <dbl>, FLT4 <dbl>, GATA3 <dbl>, IKZF1 <dbl>, KEAP1 <dbl>,
#> #   KMT2C <dbl>, KMT2D <dbl>, KRAS <dbl>, LATS1 <dbl>, MAP3K1 <dbl>, MET <dbl>,
#> #   MGA <dbl>, MLH1 <dbl>, MYC <dbl>, NCOA6 <dbl>, NCOR1 <dbl>, NF1 <dbl>,
#> #   NFE2L2 <dbl>, NOTCH1 <dbl>, NOTCH2 <dbl>, NRAS <dbl>, PASK <dbl>,
#> #   PDGFRA <dbl>, PHOX2B <dbl>, PIK3CA <dbl>, PLXNB2 <dbl>, POLE <dbl>,
#> #   PRDM1 <dbl>, PRF1 <dbl>, PTEN <dbl>, PTPRC <dbl>, RAD21 <dbl>, RASA1 <dbl>,
#> #   RB1 <dbl>, RNF43 <dbl>, SERPINB13 <dbl>, SETD2 <dbl>, SGK223 <dbl>,
#> #   SMAD4 <dbl>, SMARCA4 <dbl>, SOX2 <dbl>, SPEN <dbl>, STK11 <dbl>,
#> #   TERT <dbl>, TP53 <dbl>, TSC2 <dbl>, U2AF1 <dbl>, UBR5 <dbl>, WRN <dbl>,
#> #   WT1 <dbl>
#> 
#> $Matrix_subclonal_drivers
#> # A tibble: 99 x 80
#>    patientID   APC ARHGAP35 ARID1B ARID2 ASXL1 ATM   BAP1   BRAF CBLB  CCND1
#>    <chr>     <dbl> <chr>    <chr>  <chr> <chr> <chr> <chr> <dbl> <chr> <chr>
#>  1 CRUK0001      0 1        0      0     0     0     0         0 0     0    
#>  2 CRUK0002      0 0        0      0     0     0     0         0 0     0    
#>  3 CRUK0003      0 0        0      0     0     0     0         0 0     0    
#>  4 CRUK0004      0 0        0      0     0     0     0         0 0     0    
#>  5 CRUK0005      0 0        0      0     0     0     0         0 0     0    
#>  6 CRUK0006      0 0        0      0     0     0     0         0 0     0    
#>  7 CRUK0008      0 0        0      1     0     0     0         0 0     0    
#>  8 CRUK0009      0 0        0      0     0     0     0         0 0     0    
#>  9 CRUK0011      0 0        0      0     0     0     0         0 0     0    
#> 10 CRUK0013      0 0        0      0     0     0     0         0 0     0    
#> # … with 89 more rows, and 69 more variables: CDKN2A <dbl>, CHEK2 <dbl>,
#> #   CIC <chr>, CMTR2 <dbl>, COL2A1 <dbl>, COL5A2 <chr>, CREBBP <dbl>,
#> #   CTNNB1 <chr>, CUX1 <dbl>, CYLD <chr>, DICER1 <dbl>, DNM2 <chr>, EGFR <chr>,
#> #   EP300 <chr>, FANCC <chr>, FANCM <chr>, FAS <chr>, FAT1 <chr>, FBXW7 <dbl>,
#> #   FGFR1 <dbl>, FLT4 <chr>, GATA3 <dbl>, IKZF1 <chr>, KEAP1 <dbl>,
#> #   KMT2C <dbl>, KMT2D <chr>, KRAS <chr>, LATS1 <chr>, MAP3K1 <chr>, MET <dbl>,
#> #   MGA <chr>, MLH1 <chr>, MYC <dbl>, NCOA6 <chr>, NCOR1 <chr>, NF1 <chr>,
#> #   NFE2L2 <chr>, NOTCH1 <chr>, NOTCH2 <dbl>, NRAS <chr>, PASK <chr>,
#> #   PDGFRA <dbl>, PHOX2B <dbl>, PIK3CA <chr>, PLXNB2 <chr>, POLE <dbl>,
#> #   PRDM1 <dbl>, PRF1 <dbl>, PTEN <dbl>, PTPRC <chr>, RAD21 <dbl>, RASA1 <dbl>,
#> #   RB1 <chr>, RNF43 <chr>, SERPINB13 <dbl>, SETD2 <chr>, SGK223 <dbl>,
#> #   SMAD4 <chr>, SMARCA4 <dbl>, SOX2 <dbl>, SPEN <dbl>, STK11 <dbl>,
#> #   TERT <chr>, TP53 <chr>, TSC2 <dbl>, U2AF1 <dbl>, UBR5 <chr>, WRN <dbl>,
#> #   WT1 <dbl>
#> 
#> $Matrix_trajectories
#> # A tibble: 99 x 263
#>    patientID `ARHGAP35 --> T… `ARID1B --> ASX… `ARID1B --> COL… `ARID1B --> KRA…
#>    <chr>                <dbl>            <dbl>            <dbl>            <dbl>
#>  1 CRUK0001                 0                0                0                0
#>  2 CRUK0002                 0                0                0                0
#>  3 CRUK0003                 0                0                0                0
#>  4 CRUK0004                 0                0                0                0
#>  5 CRUK0005                 0                0                0                0
#>  6 CRUK0006                 0                0                0                0
#>  7 CRUK0007                 0                0                0                0
#>  8 CRUK0008                 0                0                0                0
#>  9 CRUK0009                 1                0                0                0
#> 10 CRUK0010                 0                0                0                0
#> # … with 89 more rows, and 258 more variables: ARID2 --> KRAS <dbl>,
#> #   ATM --> CCND1 <dbl>, ATM --> MGA <dbl>, ATM --> NCOR1 <dbl>,
#> #   BAP1 --> PIK3CA <dbl>, BAP1 --> RB1 <dbl>, BAP1 --> TP53 <dbl>,
#> #   BRAF --> TERT <dbl>, CBLB --> ARID1B <dbl>, CBLB --> DNM2 <dbl>,
#> #   CBLB --> LATS1 <dbl>, CBLB --> PTPRC <dbl>, CCND1 --> ARID1B <dbl>,
#> #   CCND1 --> FAS <dbl>, CCND1 --> IKZF1 <dbl>, CCND1 --> UBR5 <dbl>,
#> #   CDKN2A --> COL5A2 <dbl>, CDKN2A --> CTNNB1 <dbl>, CDKN2A --> KMT2D <dbl>,
#> #   CDKN2A --> NF1 <dbl>, CDKN2A --> NFE2L2 <dbl>, CDKN2A --> PTPRC <dbl>,
#> #   CDKN2A --> TP53 <dbl>, CMTR2 --> NRAS <dbl>, CMTR2 --> UBR5 <dbl>,
#> #   COL2A1 --> BAP1 <dbl>, COL2A1 --> NCOR1 <dbl>, COL5A2 --> CBLB <dbl>,
#> #   COL5A2 --> NCOR1 <dbl>, COL5A2 --> NFE2L2 <dbl>, CREBBP --> TP53 <dbl>,
#> #   CYLD --> FANCM <dbl>, DICER1 --> PLXNB2 <dbl>, DNM2 --> CBLB <dbl>,
#> #   DNM2 --> NFE2L2 <dbl>, EGFR --> ARHGAP35 <dbl>, EGFR --> CTNNB1 <dbl>,
#> #   EGFR --> RB1 <dbl>, EGFR --> TP53 <dbl>, EP300 --> CYLD <dbl>,
#> #   EP300 --> NF1 <dbl>, FANCC --> CYLD <dbl>, FANCM --> FAT1 <dbl>,
#> #   FAS --> FLT4 <dbl>, FAT1 --> ARID1B <dbl>, FAT1 --> CYLD <dbl>,
#> #   FAT1 --> DNM2 <dbl>, FAT1 --> KRAS <dbl>, FAT1 --> LATS1 <dbl>,
#> #   FAT1 --> NFE2L2 <dbl>, FAT1 --> PTPRC <dbl>, FBXW7 --> EP300 <dbl>,
#> #   FGFR1 --> COL5A2 <dbl>, FGFR1 --> KRAS <dbl>, FGFR1 --> NF1 <dbl>,
#> #   FGFR1 --> NFE2L2 <dbl>, FGFR1 --> PLXNB2 <dbl>, GL --> APC <dbl>,
#> #   GL --> ARHGAP35 <dbl>, GL --> ARID2 <dbl>, GL --> ASXL1 <dbl>,
#> #   GL --> ATM <dbl>, GL --> BAP1 <dbl>, GL --> BRAF <dbl>, GL --> CCND1 <dbl>,
#> #   GL --> CDKN2A <dbl>, GL --> CHEK2 <dbl>, GL --> CMTR2 <dbl>,
#> #   GL --> COL2A1 <dbl>, GL --> COL5A2 <dbl>, GL --> CREBBP <dbl>,
#> #   GL --> CTNNB1 <dbl>, GL --> CUX1 <dbl>, GL --> DICER1 <dbl>,
#> #   GL --> EGFR <dbl>, GL --> FANCM <dbl>, GL --> FAT1 <dbl>,
#> #   GL --> FBXW7 <dbl>, GL --> FGFR1 <dbl>, GL --> GATA3 <dbl>,
#> #   GL --> KEAP1 <dbl>, GL --> KMT2C <dbl>, GL --> KMT2D <dbl>,
#> #   GL --> KRAS <dbl>, GL --> LATS1 <dbl>, GL --> MAP3K1 <dbl>,
#> #   GL --> MET <dbl>, GL --> MGA <dbl>, GL --> MYC <dbl>, GL --> NCOA6 <dbl>,
#> #   GL --> NF1 <dbl>, GL --> NOTCH2 <dbl>, GL --> NRAS <dbl>,
#> #   GL --> PASK <dbl>, GL --> PDGFRA <dbl>, GL --> PHOX2B <dbl>,
#> #   GL --> PIK3CA <dbl>, GL --> POLE <dbl>, GL --> PRDM1 <dbl>,
#> #   GL --> PRF1 <dbl>, …
#>