Construct a REVOLVER cohort object (S3 class "rev_cohort").

Construct a REVOLVER cohort object (S3 class "rev_cohort").

revolver_cohort(dataset, CCF_parser = revolver::CCF_parser,
  ONLY.DRIVER = FALSE, MIN.CLUSTER.SIZE = 10,
  annotation = "My REVOLVER dataset")

Arguments

dataset
A dataframe in the specified format (see Online manual).

CCF_parser
A function to parse the format for the encoding of CCF or binary values for each sequenced region. A possible function is available inside REVOLVER; CCF_parser (the default of this parameter).

ONLY.DRIVER
If true, uses only annotated driver events.

MIN.CLUSTER.SIZE
Discard clusters that have less than this number of entries.

annotation
Brief cohort description.

Value

An object of class "rev_cohort" that represents a REVOLVER cohort.

See also

Examples

data(CRC)

#> Warning: data set ‘CRC’ not found

cohort = revolver_cohort(CRC, options = list(ONLY.DRIVER = FALSE, MIN.CLUSTER.SIZE = 0))

#> Error in revolver_cohort(CRC, options = list(ONLY.DRIVER = FALSE, MIN.CLUSTER.SIZE = 0)): could not find function "revolver_cohort"

# Example cohort creation with the TRACERx data
data(TRACERx_data)

#> Warning: data set ‘TRACERx_data’ not found

# To speed up the process we use only 2 patients
TRACERx_data = TRACERx_data %>%
   filter(patientID %in% c('CRUK0001', 'CRUK0002'))

#> Error in eval(lhs, parent, parent): object 'TRACERx_data' not found

cohort = revolver_cohort(TRACERx_data, annotation = 'A toy REVOLVER dataset')

#> Error in revolver_cohort(TRACERx_data, annotation = "A toy REVOLVER dataset"): could not find function "revolver_cohort"

# The S3 print/ plot for this cohort
print(cohort)

#> Error in print(cohort): object 'cohort' not found

plot(cohort)

#> Error in plot(cohort): object 'cohort' not found