Fit a REVOLVER model.

Fit a REVOLVER model.

revolver_fit(x, initial.solution = 1, max.iterations = 10, n = 10,
  ...)

Arguments

x
A REVOLVER cohort where trees per patient have been already computed.

initial.solution
Either a scalar to fix one initial condition (rank id), or NA to sample it randomly across all possivle solutions. Notice that if the inital conditin is fixed the other parameter `n` should be 1.

max.iterations
Maximum number of EM steps before forcing stop.

n
Number of initial conditions sampled to compute optimal fit.

...
Other paraemters, not used.

Value

A new object of class "rev_cohort_fit" which represents a REVOLVER cohort object with fits.

Details

The main fitting function for REVOLVER, implements the 2-steps algorithm described in the REVOLVER paper (Caravagna et al., Nature Methods volume 15, pages 707–714, 2018; https://www.nature.com/articles/s41592-018-0108-x). The output object contains a new field `$fit` which contains the fit results.

Examples

# Data released in the 'evoverse.datasets'
data('TRACERx_NEJM_2017_REVOLVER', package = 'evoverse.datasets')

new_fit = revolver_fit(TRACERx_NEJM_2017_REVOLVER)

#>  [ REVOLVER Transfer Learning fit- TRACERx NEJM 2017 ] 
#> 
#>  WARNING  Some patients have only one clone with drivers, and therefore they will just be expanded. 
#> # A tibble: 54 x 7
#>    patientID numBiopsies numMutations numDriverMutati… numClonesWithDr…
#>    <chr>           <int>        <int>            <int>            <int>
#>  1 CRUK0007            2            3                3                1
#>  2 CRUK0010            2            3                3                1
#>  3 CRUK0012            2            1                1                1
#>  4 CRUK0018            4            4                4                1
#>  5 CRUK0019            2            1                1                1
#>  6 CRUK0021            2            4                4                1
#>  7 CRUK0025            3            3                3                1
#>  8 CRUK0026            2            4                4                1
#>  9 CRUK0028            2            2                2                1
#> 10 CRUK0029            6            4                4                1
#> # … with 44 more rows, and 2 more variables: numTruncalMutations <int>,
#> #   numSubclonalMutations <int>

#> Beware: because you have set a fixed initial condition `n` will be disregarded because this EM is exhaustive.

#> 
#> Fitting  N = 99 patients 
#> 
#> 

#> Error in .:::ctree: unused argument (DataFrameToEdges)