Computation of fit confidence with the jackknife.

Computation of fit confidence with the jackknife.

revolver_jackknife(cohort, resamples = 100, leave.out = 0.1,
  cores.ratio = 0.9, options.fit = list(initial.solution = NA,
  transitive.orderings = FALSE, restarts = 0),
  options.clustering = list(use.GL = TRUE, transitive.closure = FALSE,
  min.group.size = 3, hc.method = "ward", cutoff.features_annotation = 3,
  split.method = "cutreeHybrid"), dump.partial = FALSE,
  overwrite = FALSE)

Arguments

cohort
A cohort object where fit and clusters have been computed/

resamples
Number of jackknife samples.

cores.ratio
Ratio of cores for parallel execution

options.fit
List of parameters for fitting models. See revolver_fit.

options.clustering
List of parameters for clustering with the germline node GL. See revolver_cluster.

removal
A number in [0,1] for the percentage of samples to leave out in each jackknife iteration.

Value

A cohort where a new jackknife field contains result from this analysis

Examples

if (FALSE) {
 TODO
}