Compute clusters stability via the jackknife.

For a set of clusters computed via revolver_cluster, you can compute their stability via a jackknife. routine. This funcion runs a kind of bootstrap routine where subset of patients - a desired number - is removed from the cohort and before re-computing the clusters. In this way, the co-clustering probability of each patient is computed, which leads to a mean clustering stability for each one of the original set of clusters and a frequency for the inference of a particular evolutionary trajectory.

A number of functions are available to plot the results from this jackknife analysis. Note that in general if you require a large number of runs (i.e., resamples), this computation can take some time. This implementation leverages on the easypar package to run in parallel all the re-runs, therefore we suggest to run it on a multi-core machine to appreciate a speed up in the computations.

revolver_jackknife(
  x,
  resamples = 100,
  leave.out = 0.1,
  options.fit = list(initial.solution = NA, max.iterations = 10, n = 10),
  options.clustering = list(min.group.size = 3, hc.method = "ward", split.method =
    "cutreeHybrid"),
  ...
)

Arguments

resamples
Number of jackknife samples.

options.fit
List of parameters for fitting models. See revolver_fit.

options.clustering
List of parameters for clustering with the germline node GL. See revolver_cluster.

cohort
A cohort object where fit and clusters have been computed.

removal
A number in [0,1] for the percentage of samples to leave out in each jackknife iteration.

cores.ratio
Ratio of cores for parallel execution

Value

A cohort where a new jackknife field contains result from this analysis

Examples

if (FALSE) {
 TODO
}